Reverted last WIP.

+++ client/cl_pred.c
@@ -59,7 +59,7 @@ void CL_CheckPredictionError (void)
 
 		// save for error itnerpolation
 		for (i=0 ; i<3 ; i++)
-			cl.prediction_error[i] = delta[i]*0.125f;
+			cl.prediction_error[i] = delta[i]*0.125;
 	}
 }
 
@@ -106,10 +106,10 @@ void CL_ClipMoveToEntities ( vec3_t start, vec3_t mins, vec3_t maxs, vec3_t end,
 			zd = 8*((ent->solid>>5) & 31);
 			zu = 8*((ent->solid>>10) & 63) - 32;
 
-			bmins[0] = bmins[1] = (float)-x;
-			bmaxs[0] = bmaxs[1] = (float)x;
-			bmins[2] = (float)-zd;
-			bmaxs[2] = (float)zu;
+			bmins[0] = bmins[1] = -x;
+			bmaxs[0] = bmaxs[1] = x;
+			bmins[2] = -zd;
+			bmaxs[2] = zu;
 
 			headnode = CM_HeadnodeForBox (bmins, bmaxs);
 			angles = vec3_origin;	// boxes don't rotate
@@ -238,7 +238,7 @@ void CL_PredictMovement (void)
 	pm.trace = CL_PMTrace;
 	pm.pointcontents = CL_PMpointcontents;
 
-	pm_airaccelerate = (float)atof(cl.configstrings[CS_AIRACCEL]);
+	pm_airaccelerate = atof(cl.configstrings[CS_AIRACCEL]);
 
 	pm.s = cl.frame.playerstate.pmove;
 
@@ -264,15 +264,15 @@ void CL_PredictMovement (void)
 	step = pm.s.origin[2] - oldz;
 	if (step > 63 && step < 160 && (pm.s.pm_flags & PMF_ON_GROUND) )
 	{
-		cl.predicted_step = step * 0.125f;
-		cl.predicted_step_time = (unsigned int)(cls.realtime - cls.frametime * 500);
+		cl.predicted_step = step * 0.125;
+		cl.predicted_step_time = cls.realtime - cls.frametime * 500;
 	}
 
 
 	// copy results out for rendering
-	cl.predicted_origin[0] = pm.s.origin[0]*0.125f;
-	cl.predicted_origin[1] = pm.s.origin[1]*0.125f;
-	cl.predicted_origin[2] = pm.s.origin[2]*0.125f;
+	cl.predicted_origin[0] = pm.s.origin[0]*0.125;
+	cl.predicted_origin[1] = pm.s.origin[1]*0.125;
+	cl.predicted_origin[2] = pm.s.origin[2]*0.125;
 
 	VectorCopy (pm.viewangles, cl.predicted_angles);
 }
