+++ game/g_phys.c
@@ -418,7 +418,7 @@ qboolean SV_Push (edict_t *pusher, vec3_t move, vec3_t amove)
 			temp += 0.5;
 		else
 			temp -= 0.5;
-		move[i] = 0.125f * (int)temp;
+		move[i] = 0.125 * (int)temp;
 	}
 
 	// find the bounding box
@@ -854,7 +854,7 @@ void SV_Physics_Step (edict_t *ent)
 	// friction for flying monsters that have been given vertical velocity
 	if ((ent->flags & FL_FLY) && (ent->velocity[2] != 0))
 	{
-		speed = fabsf(ent->velocity[2]);
+		speed = fabs(ent->velocity[2]);
 		control = speed < sv_stopspeed ? sv_stopspeed : speed;
 		friction = sv_friction/3;
 		newspeed = speed - (FRAMETIME * control * friction);
@@ -867,7 +867,7 @@ void SV_Physics_Step (edict_t *ent)
 	// friction for flying monsters that have been given vertical velocity
 	if ((ent->flags & FL_SWIM) && (ent->velocity[2] != 0))
 	{
-		speed = fabsf(ent->velocity[2]);
+		speed = fabs(ent->velocity[2]);
 		control = speed < sv_stopspeed ? sv_stopspeed : speed;
 		newspeed = speed - (FRAMETIME * control * sv_waterfriction * ent->waterlevel);
 		if (newspeed < 0)
@@ -884,7 +884,7 @@ void SV_Physics_Step (edict_t *ent)
 			if (!(ent->health <= 0.0 && !M_CheckBottom(ent)))
 			{
 				vel = ent->velocity;
-				speed = sqrtf(vel[0]*vel[0] +vel[1]*vel[1]);
+				speed = sqrt(vel[0]*vel[0] +vel[1]*vel[1]);
 				if (speed)
 				{
 					friction = sv_friction;
